Extract iOS review speech logic
