+++ apps/ios/Flashcards/Flashcards/Review/Speech/ReviewSpeechTypes.swift
@@ -1,0 +1,4 @@
+enum ReviewSpeechSide {
+    case front
+    case back
+}
